+++ python/GafferTractor/TractorDispatcher.py
@@ -98,7 +98,7 @@ class TractorDispatcher( GafferDispatch.Dispatcher ) :
 		# might just be member data for a subclass of one of those.
 		dispatchData = {}
 		dispatchData["scriptNode"] = rootBatch.preTasks()[0].node().scriptNode()
-		dispatchData["scriptFile"] = Gaffer.Context.current()["dispatcher:scriptFileName"]
+		dispatchData["fullScriptFile"] = Gaffer.Context.current()["dispatcher:scriptFileName"]
 		dispatchData["batchesToTasks"] = {}
 		dispatchData["taskData"] = {}
 
@@ -180,7 +180,7 @@ class TractorDispatcher( GafferDispatch.Dispatcher ) :
 
 			args = [
 				"gaffer", "execute",
-				"-script", dispatchData["scriptFile"],
+				"-script", batch.blindData()["dispatcher:scriptFileName"] if "dispatcher:scriptFileName" in batch.blindData() else dispatchData["fullScriptFile"],
 				"-nodes", nodeName,
 				"-frames", frames,
 			]